[FIX] Nomenclatura de contrato em abate e correção de tipagem dos ids nos contratos para delete, e alteração na tipagem de FazendaID para int

+++ application/domain/abate-domain.go
@@ -23,7 +23,7 @@ type DistribuicaoPeso struct {
 
 type DadosGeraisAbate struct {
 	DataAbate        string
-	FazendaID        uint64
+	FazendaID        int
 	NomeFrigorifico  string
 	CategoriaAnimal  string
 	PrecoFunrural    float64
